feat: Support for managing the service to service KMS authorization policy has been removed from this module. For more infomation on why, see [About KMS encryption](https://github.com/terraform-ibm-modules/terraform-ibm-event-streams/tree/main?tab=readme-ov-file#about-kms-encryption). This now means that the auth policy must be in place before you attempt to provision an Event Streams instance with KMS encryption enabled.<br>* Due to this change, the following variables have bee removed `skip_iam_authorization_policy`, `existing_kms_instance_guid`, and `kms_encryption_enabled`. If you wish to enable KMS encryption, just pass a value for the `kms_key_crn` input variable. (#178)
BREAKING CHANGE: If updating to this version from a previous version, and if you were using the module to manage the KMS auth policy, you will see the auth policy as identified for destroy. In order to ensure there is no interruption to service when updating, it is recommended that you create a global auth policy (outside of this terraform state) to allow all Event Stream instances in your account read access to the KMS service you are using before you proceed with upgrade. Once your upgrade is complete, and terraform has destroyed the old auth policy, you may wish to create a new instance-to-instance scoped policy if you wish to remove the globally scoped one. It is important not to create the policy with Terraform if it shares the same state as this module. Refer to [About KMS encryption](https://github.com/terraform-ibm-modules/terraform-ibm-event-streams/tree/main?tab=readme-ov-file#about-kms-encryption) for more details.